backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3
redis==5.2.0
openai==1.44.0
PyJWT==2.10.1
loguru==0.7.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.12
cryptography==43.0.3
rq==1.16.2
